try to fix epoch size

+++ train.py
@@ -100,9 +100,9 @@ model.compile(loss='mae', optimizer='adam')
 print(len(train_samples))
 print(len(lines))
 model.fit_generator(train_generator, 
-    samples_per_epoch= int(np.ceil(len(train_samples)/BATCH_SIZE)), 
+    samples_per_epoch= int(np.ceil(len(train_samples)/BATCH_SIZE).astype( np.int32 ), 
     validation_data=validation_generator, 
-    nb_val_samples=int(np.ceil(len(validation_samples)/BATCH_SIZE)), 
+    nb_val_samples=np.ceil(len(validation_samples)/BATCH_SIZE).astype(np.int32), 
     nb_epoch=EPOCHS)
 
 model.save('model.h5')
